Allow Surface to reference an external window object
This adds a new trait, WindowAbstract, which window-management
crates can implement to allow the a Surface to keep its associated
window alive.

All of the surface-creation fns have been updated to take an
Option<Arc<WindowAbstract>>, which is simply stored in the
Surface to keep a reference.

This also updates vulkano-win to use a small wrapper around
winit::Window, in order to implement the WindowAbstract trait.

This fixes #844