Allow Surface to ref an external window object
This fixes #844

This adds a new trait, WindowAbstract, which window-management
crates can implement to allow the Surface to keep its associated
window alive.

Window management crates can optionally set a window after the
Surface is created. Because Surfaces are always Arcs, this
involves a smattering of additional `unsafe`. This unsafe
allows us to implement this without any breaking API changes.