Allow CpuAccessibleBuffer to take a custom memory Pool
Previously, CpuAccessibleBuffer would always use the StdMemoryPool
that the Device provides. This adds new methods, allowing it to be
constructed with a custom Pool instead. The type of the buffers
returned from these new constructors may be different, if the memory
pool has a different associated allocation type.

This is a slightly breaking change, since CpuAccessibleBuffer::raw
now takes an additional parameter.